Updated the code to reflect compatibility with Ar
Updated code to reflect compatibility with Arduino Uno. Added a check statement to ensure the Buffer length is at least 53 bytes. The Uno's ram is not as large as the Teensy. The main cause of this problem were the use of print statements. Those were changed to allow less usage of the SRAM. The alignment of the structs and union were redone to reflect quick reading and writing.